refact: refactoring structs to float64

+++ model/handler/request/RequestApproverHandler.go
@@ -1,6 +1,6 @@
 package model
 
 type RequestApproverHandler struct {
-	ClientId string `json:"clientId"`
-	Amount   int64  `json:"amount"`
+	ClientId string  `json:"clientId"`
+	Amount   float64 `json:"amount"`
 }
